requirements.txt
aiogram
aiohttp
openai
python-dotenv